Move ViewRegistry + Helper into internal package
