+++ sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/aggregator/AbstractSumAggregator.java
@@ -34,10 +34,10 @@ abstract class AbstractSumAggregator<T> extends AbstractAggregator<T> {
 
   /**
    * Resolve whether the aggregator should be stateful. For the special case {@link
-   * InstrumentType#OBSERVABLE_SUM} and {@link InstrumentType#OBSERVABLE_UP_DOWN_SUM} instruments, state
-   * is required if temporality is {@link AggregationTemporality#DELTA}. Because the observed values
-   * are cumulative sums, we must maintain state to compute delta sums between collections. For
-   * other instruments, state is required if temporality is {@link
+   * InstrumentType#OBSERVABLE_SUM} and {@link InstrumentType#OBSERVABLE_UP_DOWN_SUM} instruments,
+   * state is required if temporality is {@link AggregationTemporality#DELTA}. Because the observed
+   * values are cumulative sums, we must maintain state to compute delta sums between collections.
+   * For other instruments, state is required if temporality is {@link
    * AggregationTemporality#CUMULATIVE}.
    *
    * @param instrumentType the instrument type
@@ -57,8 +57,9 @@ abstract class AbstractSumAggregator<T> extends AbstractAggregator<T> {
   /**
    * Resolve the aggregator merge strategy. The merge strategy is SUM in all cases except where
    * temporality is {@link AggregationTemporality#DELTA} and instrument type is {@link
-   * InstrumentType#OBSERVABLE_SUM} or {@link InstrumentType#OBSERVABLE_UP_DOWN_SUM}. In these special
-   * cases, the observed values are cumulative sums so we must take a diff to compute the delta sum.
+   * InstrumentType#OBSERVABLE_SUM} or {@link InstrumentType#OBSERVABLE_UP_DOWN_SUM}. In these
+   * special cases, the observed values are cumulative sums so we must take a diff to compute the
+   * delta sum.
    *
    * @param instrumentType the instrument type
    * @param temporality the temporality
