+++ sdk/metrics/src/test/java/io/opentelemetry/sdk/metrics/aggregator/DoubleMinMaxSumCountAggregatorTest.java
@@ -31,11 +31,7 @@ class DoubleMinMaxSumCountAggregatorTest {
           Resource.getDefault(),
           InstrumentationLibraryInfo.empty(),
           InstrumentDescriptor.create(
-              "name",
-              "description",
-              "unit",
-              InstrumentType.HISTOGRAM,
-              InstrumentValueType.DOUBLE));
+              "name", "description", "unit", InstrumentType.HISTOGRAM, InstrumentValueType.DOUBLE));
 
   @Test
   void createHandle() {
